Move ValidationRules to Input component
