+++ typings/commonTypes.ts
@@ -4,10 +4,3 @@ export enum RoutesList {
   register = '/register',
   library = '/library',
 }
-
-export type ValidationRules = {
-  minSymbols?: number;
-  phone?: boolean;
-  email?: boolean;
-  isRequired?: boolean;
-};
